Conflicts:
	cocos2d-mac.xcodeproj/project.pbxproj

Former-commit-id: ab514c5e7aab797e8423671d7cb0a2c71567cb04

+++ cocos2d/CCActionInstant.m
@@ -248,10 +248,10 @@
 
 -(NSString*) description
 {
-	return [NSString stringWithFormat:@"<%@ = %p | Tag = %i | selector = %@>",
+	return [NSString stringWithFormat:@"<%@ = %p | Tag = %ld | selector = %@>",
 			[self class],
 			self,
-			tag_,
+			(long)tag_,
 			NSStringFromSelector(selector_)
 			];
 }
